+++ crates/dav/src/card/copy_move.rs
@@ -25,12 +25,13 @@ use hyper::StatusCode;
 use store::write::BatchBuilder;
 use store::{
     ValueKey,
-    write::{AlignedBytes, Archive},
+    write::{AlignedBytes, Archive, ValueClass},
 };
 use trc::AddContext;
 use types::{
     acl::Acl,
     collection::{Collection, SyncCollection, VanishedCollection},
+    field::PrincipalField,
 };
 
 pub(crate) trait CardCopyMoveRequestHandler: Sync + Send {
@@ -826,6 +827,25 @@ async fn copy_container(
                 &mut batch,
             )
             .caused_by(trc::location!())?;
+
+        // Reset default address book id
+        let default_book_id = server
+            .store()
+            .get_value::<u32>(ValueKey {
+                account_id: from_account_id,
+                collection: Collection::Principal.into(),
+                document_id: 0,
+                class: ValueClass::Property(PrincipalField::DefaultAddressBookId.into()),
+            })
+            .await
+            .caused_by(trc::location!())?;
+        if default_book_id.is_some_and(|id| id == from_document_id) {
+            batch
+                .with_account_id(from_account_id)
+                .with_collection(Collection::Principal)
+                .with_document(0)
+                .clear(PrincipalField::DefaultAddressBookId);
+        }
     }
 
     let preference = book.preferences.into_iter().next().unwrap();