+++ CHANGELOG.md
@@ -24,6 +24,7 @@ If you are upgrading from v0.16.x, replace the binary (or run `docker pull`). If
 ## Fixed
 - S3: Fix outdated upstream `af-south-1` region configuration.
 - Setup wizard: SQL directories set to use the main data store are now validated against the data store being configured.
+- CardDAV: Delete default address book id when deleting the default address book.
 
 ## [0.16.19] - 2026-08-24